+++ testing/wasm-helpers/src/lib.rs
@@ -17,6 +17,7 @@ macro_rules! build_configured_canister {
             .with_sql_ddl_enabled(config.canister_sql_ddl_enabled($canister_name))
             .with_sql_fixtures_enabled(config.canister_sql_fixtures_enabled($canister_name))
             .with_metrics_enabled(config.canister_metrics_enabled($canister_name))
+            .with_metrics_extended_enabled(config.canister_metrics_extended_enabled($canister_name))
             .with_metrics_reset_enabled(config.canister_metrics_reset_enabled($canister_name))
             .with_snapshot_enabled(config.canister_snapshot_enabled($canister_name))
             .with_schema_enabled(config.canister_schema_enabled($canister_name));